% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alteplase}
\alias{alteplase}
\title{Effect of time to treatment on efficacy of rt-PA in treating acute stroke}
\format{A data frame with 3669 rows and 3 variables:
\describe{
  \item{time}{Time interval (minutes) between stroke onset and treatment}
  \item{treat}{Type of treatment received}
  \item{mRS}{Outcome at 3 months measured using the modified Rankin Scale}
}}
\usage{
alteplase
}
\description{
A dataset reconstructing reported outcomes after stroke
as reported in Lees et. al. (2010) which pooled the results
of several studies.
}
\references{
Lees, K. R., Bluhmki, E., Von Kummer, R., Brott, T. G.,
Toni, D., Grotta, J. C.,  Albers, G. W., Kaste, M., Marler, J. R.,
Hamilton, S. A., Tilley, B. C., Davis, S. M., Donnan, G. A., Hacke, W.
 (2010).
Time to treatment with intravenous alteplase and outcome in stroke:
an updated pooled analysis of ECASS, ATLANTIS, NINDS, and EPITHET trials.
\emph{The Lancet}, 375(9727), 1695-1703.
}
\keyword{datasets}
